requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py==2.5.0a5125+g7db879b5
mysql-connector-python==9.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
